+++ metadata.rb
@@ -1,11 +1,11 @@
 name 'slack_handler'
 
-maintainer       'Rackspace'
-maintainer_email 'rackspace-cookbooks@rackspace.com'
+maintainer       'Josh Gitlin'
+maintainer_email 'jgitlin@pinnacle21.com'
 license          'Apache-2.0'
 description      'Installs/Configures a Chef handler for reporting results to a Slack channel.'
 
-version          '1.0.0'
+version          '1.1.0'
 
 supports 'ubuntu'
 supports 'centos'
@@ -14,6 +14,6 @@ supports 'redhat'
 supports 'debian'
 supports 'windows'
 
-source_url 'https://github.com/rackspace-cookbooks/chef-slack_handler'
-issues_url 'https://github.com/rackspace-cookbooks/chef-slack_handler/issues'
+source_url 'https://github.com/jgitlin-p21/chef-slack_handler'
+issues_url 'https://github.com/jgitlin-p21/chef-slack_handler/issues'
 chef_version '>= 14.0'